Implement Test route /run
Running the tests from the frontend requires having a route on the
server to do so. The route /run has been made that executes the tests
and returns the result to the requester. Once the API has a route to
store this information it will also send the information to the API.

Currently there is one test case that verifies the manager can login and
the route goes to /overview once login is complete. When more tests are
added in the parsing code will need to be updated as the structure will
likely change. If more suites are added then the entire shape of the
JSON should be updated.

Current response shape is:
{
    status: 1 // Integer 1 -> pass, 0 -> fail
    tests : [
        {
	    // See Test definition in frontend/server/models/Test.js
	}
    ]
}