Integrate API and Production (#23)
* Ignored the config for hibernate

Config file has sensitive information that will need to be transfered to
server manually.

* Integrated the node server with API

The proxy will send /api routes to Tomcat. The route to hit the project2
api will be /api/project2/{route for api request}. This will be prepped
for deployement onto server.

* Add route to server angular application

Routes will now go to angular application for SPA.

* Linux driver for Selinium added

* Attempting listing path to config absolute

* Server Configurations for Producation

* Added the Constants file for defining system dependant information

* Integrate Constants with new tests

Moves constants to be defined in one location to simplify the build
process on the server. This file should be ignored in the furture to
avoid each developers local changes from constantly changing the file.

* Send batch data to database

Sends data straight to the database then once entered sends to the
client with the inserted ids for the tests.

* Fix the error refreshing maven test

The maven test was failing but error silencing handled the error and
returned the outdated report.

* README file update with paths to constants